another shit happens

+++ src/engine/physics_material.cpp
@@ -2,15 +2,15 @@
 
 using namespace ark;
 
-static material::data material_data[static_cast<size_t>(material::type::out_of)] = {};
+static material::data material_data[static_cast<uint8_t>(material::type::out_of)] = {};
 
 void
 material::init()
 {
-	material_data[static_cast<size_t>(type::rubber)].friction = 0.1f;
-	material_data[static_cast<size_t>(type::rubber)].density = 1.f;
-	material_data[static_cast<size_t>(type::rubber)].restitution = 0.9f;
-}
+	material_data[static_cast<uint8_t>(type::rubber)].friction = 0.1f;
+	material_data[static_cast<uint8_t>(type::rubber)].density = 1.f;
+	material_data[static_cast<uint8_t>(type::rubber)].restitution = 0.9f;
+}							  
 
 void
 material::destroy()
@@ -20,5 +20,5 @@ material::destroy()
 const material::data& 
 material::get(type type)
 {
-	return material_data[static_cast<size_t>(type)];
+	return material_data[static_cast<uint8_t>(type)];
 }
